[SPARK-44205][SQL] Extract Catalyst Code from DecimalType
