+++ sql/catalyst/src/main/scala/org/apache/spark/sql/types/DataTypeExpression.scala
@@ -65,3 +65,14 @@ object IntegralTypeExpression {
     e.dataType.isInstanceOf[IntegralType]
   }
 }
+
+object DecimalExtractor {
+  def unapply(e: Expression): Option[(Int, Int)] = e.dataType match {
+    case t: DecimalType => Some((t.precision, t.scale))
+    case _ => None
+  }
+}
+
+object DecimalExpression {
+  def unapply(e: Expression): Boolean = e.dataType.isInstanceOf[DecimalType]
+}
